+++ packages/jupyter/src/output.tsx
@@ -1,13 +1,11 @@
 import type { GenericNode } from 'myst-common';
 import { KnownCellOutputMimeTypes } from 'nbtx';
 import type { MinifiedMimeOutput, MinifiedOutput } from 'nbtx';
-import classNames from 'classnames';
-import { SafeOutputs } from './safe.js';
-import { JupyterOutputs } from './jupyter.js';
+import { SafeOutput } from './safe.js';
+import { JupyterOutput } from './jupyter.js';
 import { useMemo } from 'react';
 import { useCellExecution } from './execute/index.js';
-import { usePlaceholder } from './decoration.js';
-import { Details, MyST } from 'myst-to-react';
+import { useOutputsContext } from './providers.js';
 
 export const DIRECT_OUTPUT_TYPES = new Set(['stream', 'error']);
 
@@ -19,89 +17,31 @@ export const DIRECT_MIME_TYPES = new Set([
   KnownCellOutputMimeTypes.ImageBmp,
 ]) as Set<string>;
 
-export function allOutputsAreSafe(
-  outputs: MinifiedOutput[],
+export function isOutputSafe(
+  output: MinifiedOutput,
   directOutputTypes: Set<string>,
   directMimeTypes: Set<string>,
 ) {
-  if (!outputs || outputs.length === 0) return true;
-  return outputs.reduce((flag, output) => {
-    if (directOutputTypes.has(output.output_type)) return flag && true;
-    const data = (output as MinifiedMimeOutput).data;
-    const mimetypes = data ? Object.keys(data) : [];
-    const safe =
-      'data' in output &&
-      Boolean(output.data) &&
-      mimetypes.every((mimetype) => directMimeTypes.has(mimetype));
-    return flag && safe;
-  }, true);
-}
-
-export function JupyterOutput({
-  outputId,
-  identifier,
-  data,
-  align,
-  className,
-}: {
-  outputId: string;
-  identifier?: string;
-  data: MinifiedOutput[];
-  align?: 'left' | 'center' | 'right';
-  className?: string;
-}) {
-  const { ready } = useCellExecution(outputId);
-  const outputs: MinifiedOutput[] = data;
-  const allSafe = useMemo(
-    () => allOutputsAreSafe(outputs, DIRECT_OUTPUT_TYPES, DIRECT_MIME_TYPES),
-    [outputs],
-  );
-  const placeholder = usePlaceholder();
-
-  let component;
-  if (allSafe && !ready) {
-    if (placeholder && (!outputs || outputs.length === 0)) {
-      if (placeholder) {
-        return <MyST ast={placeholder} />;
-      }
-    }
-    component = <SafeOutputs keyStub={outputId} outputs={outputs} />;
-  } else {
-    component = <JupyterOutputs id={outputId} outputs={outputs} />;
-  }
-
+  if (directOutputTypes.has(output.output_type)) return true;
+  const data = (output as MinifiedMimeOutput).data;
+  const mimetypes = data ? Object.keys(data) : [];
   return (
-    <div
-      id={identifier || undefined}
-      data-mdast-node-id={outputId}
-      className={classNames(
-        'max-w-full overflow-y-visible overflow-x-auto m-0 group not-prose relative',
-        {
-          'text-left': !align || align === 'left',
-          'text-center': align === 'center',
-          'text-right': align === 'right',
-          'mb-5': outputs && outputs.length > 0,
-        },
-        className,
-      )}
-    >
-      {component}
-    </div>
+    'data' in output &&
+    Boolean(output.data) &&
+    mimetypes.every((mimetype) => directMimeTypes.has(mimetype))
   );
 }
 
 export function Output({ node }: { node: GenericNode }) {
-  const output = (
-    <JupyterOutput
-      className={classNames({ hidden: node.visibility === 'remove' })}
-      outputId={node.id}
-      identifier={node.identifier}
-      align={node.align}
-      data={node.data}
-    />
-  );
-  if (node.visibility === 'hide') {
-    return <Details title="Output">{output}</Details>;
+  const { outputsId } = useOutputsContext();
+  const { ready } = useCellExecution(outputsId);
+  // FUTURE: we'll be rendering AST outputs directly in future
+  const maybeSafeOutput = useMemo(() => node.jupyter_data, [node]);
+  const isSafe = isOutputSafe(maybeSafeOutput, DIRECT_OUTPUT_TYPES, DIRECT_MIME_TYPES);
+
+  if (isSafe && !ready) {
+    return <SafeOutput output={maybeSafeOutput} />;
   }
-  return output;
+  // TODO: myst-jp-output should be added to the first child div
+  return <JupyterOutput outputsId={outputsId} output={maybeSafeOutput} />;
 }