+++ packages/frontmatter/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @myst-theme/frontmatter
 
+## 1.0.0
+
+## 0.18.0
+
+### Minor Changes
+
+- dee582b: Add human-readable CSS classes to myst components
+- c6699ff: Add hide_authors handling
+
+### Patch Changes
+
+- 10a7355: remove Jupyter Logo that is not clickable
+
 ## 0.17.1
 
 ## 0.17.0