+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @myst-theme/icons
 
+## 1.0.0
+
+## 0.18.0
+
+### Minor Changes
+
+- dee582b: Add human-readable CSS classes to myst components
+
 ## 0.17.1
 
 ## 0.17.0