+++ packages/jupyter/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @myst-theme/jupyter
 
+## 1.0.0
+
+### Major Changes
+
+- 16a3bc8: Updates the theme to accept a new AST structure for notebook output. These changes are not backwards compatible, and after this change, only the new AST structure is supported. Meaning, content upgrades are required to use the new themes. See https://github.com/jupyter-book/mystmd/pull/1903 for related AST changes.
+
+### Patch Changes
+
+- Updated dependencies [bd19488]
+  - myst-to-react@1.0.0
+  - @myst-theme/providers@1.0.0
+
+## 0.18.0
+
+### Minor Changes
+
+- dee582b: Add human-readable CSS classes to myst components
+
+### Patch Changes
+
+- ef1b152: Fix execution button positioning
+- Updated dependencies [60cca10]
+- Updated dependencies [49fc460]
+  - @myst-theme/providers@0.18.0
+  - myst-to-react@0.18.0
+
 ## 0.17.1
 
 ### Patch Changes