+++ themes/article/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @myst-theme/article
 
+## 1.0.0
+
+### Major Changes
+
+- 16a3bc8: Updates the theme to accept a new AST structure for notebook output. These changes are not backwards compatible, and after this change, only the new AST structure is supported. Meaning, content upgrades are required to use the new themes. See https://github.com/jupyter-book/mystmd/pull/1903 for related AST changes.
+
+### Patch Changes
+
+- Updated dependencies [16a3bc8]
+- Updated dependencies [bd19488]
+  - @myst-theme/jupyter@1.0.0
+  - @myst-theme/common@1.0.0
+  - @myst-theme/styles@1.0.0
+  - @myst-theme/providers@1.0.0
+  - @myst-theme/site@1.0.0
+  - @myst-theme/icons@1.0.0
+
+## 0.18.0
+
+### Minor Changes
+
+- dee582b: Add human-readable CSS classes to myst components
+
+### Patch Changes
+
+- c6699ff: Add hide_authors handling
+- Updated dependencies [dee582b]
+- Updated dependencies [3a334e5]
+- Updated dependencies [c6699ff]
+- Updated dependencies [b5a3495]
+- Updated dependencies [ef1b152]
+- Updated dependencies [60cca10]
+- Updated dependencies [7190e6c]
+  - @myst-theme/jupyter@0.18.0
+  - @myst-theme/common@0.18.0
+  - @myst-theme/icons@0.18.0
+  - @myst-theme/site@0.18.0
+  - @myst-theme/providers@0.18.0
+  - @myst-theme/styles@0.18.0
+
 ## 0.17.1
 
 ### Patch Changes
@@ -898,7 +938,6 @@
 ### Patch Changes
 
 - 0536fa7: Rename styles to allow for additional grid systems:
-
   - `article-grid-gap` to `grid-gap`
   - `article-subgrid-gap` to `subgrid-gap`